requirements.txt
pydantic==2.11.7
fastapi==0.116.1
uvicorn==0.35.0
GitPython==3.1.44
llama-cpp-python==0.3.13
pydantic-settings==2.10.1
requests==2.32.4
python-dotenv==1.0.1
ipykernel==6.29.5
langchain==0.3.26
langchain-community==0.3.27 
httpx==0.28.1
beautifulsoup4==4.13.4
google-cloud-aiplatform==1.59.0
pytest==8.4.1
redis==6.2.0



